Rename namespace of *TestCase classes
