+++ composer.json
@@ -38,7 +38,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Respect\\Validation\\Test\\": "tests/library/"
+            "Respect\\Validation\\": "tests/library/"
         }
     },
     "extra": {
