+++ tests/unit/Rules/CountableTest.php
@@ -11,8 +11,6 @@
 
 namespace Respect\Validation\Rules;
 
-use Respect\Validation\Test\RuleTestCase;
-
 /**
  * @group  rule
  * @covers Respect\Validation\Rules\Countable
